Sign CI builds with throwaway self-signed cert

+++ ps1/Publish-Release.ps1
@@ -3,6 +3,7 @@ $ProgressPreference = 'SilentlyContinue'
 Add-Type -AssemblyName System.IO.Compression, System.IO.Compression.FileSystem
 
 $root = Split-Path -Path $PSScriptRoot -Parent
+Write-Host "Root: $root"
 
 $publishDir = "$root\publish"
 if (Test-Path $publishDir) { [System.IO.Directory]::Delete($publishDir, $true) | Out-Null }
@@ -13,23 +14,38 @@ $buildDir = "$root\publish\build"
 $downloadsDir = "$root\downloads"
 [System.IO.Directory]::CreateDirectory($downloadsDir) | Out-Null
 
-function Publish-PlatformRelease
+# Windows SDK
+$windowsSdkBaseDir = "C:\Program Files (x86)\Windows Kits\10\Redist"
+$windowsSdkVersion = `
+    Get-ChildItem -Path $windowsSdkBaseDir | 
+    Where-Object { $_.Name -match '^10\.0\.\d+\.\d+$' } | 
+    Sort-Object Name -Descending | 
+    Select-Object -First 1 -ExpandProperty Name
+
+$makeappx = "C:\Program Files (x86)\Windows Kits\10\bin\$windowsSdkVersion\x64\makeappx.exe"
+if (Test-Path $makeappx) {
+    Write-Output "MakeAppx: $makeappx"
+} else {
+	throw "MakeAppx not found!"
+}
+
+function Publish-App
 {
 	param
 	(
-		[Parameter(Mandatory = $true)]
-		[string] $Arch
+		[Parameter(Mandatory = $true)] [string] $Arch
 	)
 
 	Write-Host "Publishing $Arch."
 	$dir = "$buildDir\$Arch"
 	[System.IO.Directory]::CreateDirectory($dir) | Out-Null
-	dotnet publish "$root/src/J.App/J.App.csproj" --output "$dir" --self-contained --runtime "win-$Arch" --configuration Release
+	dotnet publish "$root/src/J.App/J.App.csproj" --output "$dir" --self-contained --runtime "win-$Arch" --configuration Release --verbosity quiet
+	Remove-Item -Path "$dir\*.pdb" -Force
 }
 
 function Publish-Launcher
 {
-	dotnet publish "$root/src/J.Launcher/J.Launcher.csproj" --output "$buildDir" --self-contained --runtime "win-x64" --configuration Release
+	dotnet publish "$root/src/J.Launcher/J.Launcher.csproj" --output "$buildDir" --self-contained --runtime "win-x64" --configuration Release --verbosity quiet
 	Remove-Item -Path "$buildDir\*.pdb" -Force
 }
 
@@ -41,6 +57,9 @@ function Get-FfmpegX64
 	{
 		Write-Host "Downloading ffmpeg/x64."
 		& curl.exe -Lso "$zipFilePath" "$url" | Out-Host
+		if ($LastExitCode -ne 0) {
+			throw "Failed to download ffmpeg/x64."
+		}
 	}
 
 	Write-Host "Extracting ffmpeg/x64."
@@ -66,6 +85,9 @@ function Get-FfmpegArm64
 	{
 		Write-Host "Downloading ffmpeg/arm64."
 		& curl.exe -Lso "$zipFilePath" "$url" | Out-Host
+		if ($LastExitCode -ne 0) {
+			throw "Failed to download ffmpeg/arm64."
+		}
 	}
 
 	Write-Host "Extracting ffmpeg/arm64."
@@ -82,6 +104,9 @@ function Get-Vlc
 	{
 		Write-Host "Finding current release of vlc/x64."
 		& curl.exe -Lso "$htmlFilePath" "$url" | Out-Host
+		if ($LastExitCode -ne 0) {
+			throw "Failed to download vlc.html."
+		}
 	}
 
 	# Read the html file, find the first line that contains ".zip"
@@ -97,6 +122,9 @@ function Get-Vlc
 	{
 		Write-Host "Downloading vlc/x64."
 		& curl.exe -Lso "$zipFilePath" "$url" | Out-Host
+		if ($LastExitCode -ne 0) {
+			throw "Failed to download vlc/x64."
+		}
 	}
 
 	Write-Host "Extracting vlc/x64."
@@ -106,6 +134,11 @@ function Get-Vlc
 
 	# Get the only subdirectory of $dstDir.
 	$subdir = Get-ChildItem -Path $dstDir -Directory | Select-Object -First 1
+
+	# Confirm $subdir is the directory with vlc.exe in it.
+	if (-not (Test-Path "$($subdir.FullName)\vlc.exe")) {
+		throw "Failed to find vlc.exe in extracted directory."
+	}
 
 	# Move everything in $subdir to $dstDir.
 	Move-Item -Path "$($subdir.FullName)\*" -Destination $dstDir -Force
@@ -131,16 +164,56 @@ function New-Msix
 	Write-Host "Creating MSIX package."
 	$msixFilePath = "$root\publish\Jackpot.msix"
 	if (Test-Path $msixFilePath) { Remove-Item -Path $msixFilePath -Force }
-	& "MakeAppx.exe" pack /d "$buildDir" /p "$msixFilePath"
+	& "$makeappx" pack /d "$buildDir" /p "$msixFilePath"
+	if ($LastExitCode -ne 0) {
+		throw "Failed to create MSIX package."
+	}
+}
 
-	# .msix file still needs to be signed at this point. That will be done offline for release builds.
+function Add-TemporarySignature
+{
+	Write-Host "Signing MSIX package with self-signed certificate."
+
+	$cert = New-SelfSignedCertificate `
+		-Type Custom `
+		-Subject "CN=Brian Luft" `
+		-KeyUsage DigitalSignature `
+		-KeyAlgorithm RSA `
+		-KeyLength 2048 `
+		-CertStoreLocation "Cert:\CurrentUser\My" `
+		-NotAfter (Get-Date).AddDays(7) `
+		-FriendlyName "Temporary CI Build Certificate" `
+		-TextExtension @("2.5.29.19={text}CA=false&pathlength=0")
+
+	try
+	{
+		# Export to .pfx (public key + private key).
+		$password = ConvertTo-SecureString -String "password" -Force -AsPlainText
+		Export-PfxCertificate -Cert $cert -FilePath "$publishDir\private.pfx" -Password $password | Out-Null
+
+		# Export to .cer (public key only).
+		Export-Certificate -Cert $cert -FilePath "$publishDir\ci-certificate.cer" | Out-Null
+
+		# Sign the .msix.
+		& "$root\ps1\Add-MsixSignature.ps1" -CertificatePath "$publishDir\private.pfx" -CertificatePassword "password"
+	}
+	finally
+	{
+		# Destroy the private certificate.
+		Remove-Item -Path $cert.PSPath
+
+		if (Test-Path "$publishDir\private.pfx") {
+			Remove-Item -Path "$publishDir\private.pfx" -Force
+		}
+	}
 }
 
 Copy-MiscFiles
 Publish-Launcher
-Publish-PlatformRelease -Arch "x64"
-Publish-PlatformRelease -Arch "arm64"
+Publish-App -Arch "x64"
+Publish-App -Arch "arm64"
 Get-FfmpegX64
 Get-FfmpegArm64
 Get-Vlc
 New-Msix
+Add-TemporarySignature
